Corrected deleting images request

+++ frontend/app/api/images/delete/route.ts
@@ -13,21 +13,20 @@ export async function DELETE(
     return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
   }
 
-  const formData = await req.formData();
-  const URLs = formData.getAll("URLs") as string[];
+  const { URLs } = await req.json();
+  if (!URLs || !Array.isArray(URLs)) {
+    return NextResponse.json({ error: "Invalid input: URLs should be an array" }, { status: 400 });
+  }
+  // const URLs = formData.getAll("URLs") as string[];
 
   // Process each URL individually
-  const results = [];
 
   for (const url of URLs) {
-    const response = await fetchWithRefresh(`${API_URL}/image/?url=${url}`, {
+    await fetchWithRefresh(`${API_URL}/image/?url=${url}`, {
       method: "DELETE",
     });
-
-    const result = await response.json();
-    results.push(result);
   }
 
   // Return combined results
-  return NextResponse.json({ results }, { status: 200 });
+  return NextResponse.json({}, { status: 200 });
 }
